src/deprecated/requirements.txt
matplotlib
pandas
liac-arff
scikit-learn
PyGObject
